Move code into index/ and add test indices
